added wating list

+++ Server/src/mysql/MySQLConnection.java
@@ -546,14 +546,15 @@ public class MySQLConnection {
 				return "Subscriber Number" + order.getId() + " " + "is not a guide";
 			order.setId(subscriber.getID());
 		}
-		Park park=(Park)getCertainPark(order.getParkName());
-		if(park==null)
+		Park park = (Park) getCertainPark(order.getParkName());
+		if (park == null)
 			return null;
-		if (park.getParkCurrentVisitors()+order.getNumOfVisitors()<park.getParkMaxVisitorsDefault()) {
+		if (park.getParkCurrentVisitors() + order.getNumOfVisitors() < park.getParkMaxVisitorsDefault()) {
 			return insertNewOrder(order, OrderStatus.APPROVED, true, new Boolean(false));
 		}
 		return null;
 	}
+
 	public static List<Order> getUnfinishedOrdersById(String id) throws SQLException {
 		List<Order> orders = new ArrayList<Order>();
 		String query = "Select * From orders where id_fk=? AND (status='WAITING' OR status='PENDING_APPROVAL_FROM_WAITING_LIST' OR status='ACTIVE' OR status='PENDING_FINAL_APPROVAL');";
@@ -569,7 +570,14 @@ public class MySQLConnection {
 		return orders;
 	}
 
-	public static Boolean changeOrderStatus(String orderNum, OrderStatus newStatus) throws SQLException {
+	public static Boolean changeOrderStatus(String orderNum, OrderStatus newStatus)
+			throws SQLException, NumberFormatException, ParseException {
+		Order order = getCertainOrder(orderNum);
+		if (order == null || order.getStatus().equals(OrderStatus.CANCELLED)
+				|| order.getStatus().equals(OrderStatus.EXPIRED)) {
+			return false;
+		}
+		System.out.println(order.getStatus().name() + " " + newStatus.name());
 		String query = "Update orders SET status=? WHERE orderNum=?";
 		String queryDelSms = "DELETE FROM smsSend WHERE orderNum_fk=?";
 		PreparedStatement delSms = con.prepareStatement(queryDelSms);
@@ -579,10 +587,39 @@ public class MySQLConnection {
 		approveOrder.setString(2, orderNum);
 		approveOrder.executeUpdate();
 		delSms.executeUpdate();
+		if (newStatus.equals(OrderStatus.CANCELLED) && !order.getStatus().equals(OrderStatus.WAITING)) {
+
+			String parkName = order.getParkName();
+			String dateOfOrder = order.getDateOfOrder();
+			String timeOfOrder = order.getTimeOfOrder();
+			Map<String, Map<String, List<String>>> checkWating = new LinkedHashMap<String, Map<String, List<String>>>();
+			Map<String, List<String>> dateAndTime = new LinkedHashMap<String, List<String>>();
+			List<String> times = new ArrayList<String>();
+			times.add(timeOfOrder);
+			dateAndTime.put(dateOfOrder, times);
+			checkWating.put(parkName, dateAndTime);
+			checkWatingList(checkWating);
+
+		}
 		return true;
 	}
 
-	public static Order activateOrderFromWatingList(Order order) throws SQLException {
+	private static Order getCertainOrder(String orderNum) throws SQLException {
+		String query = "Select * From orders where orderNum=?";
+		PreparedStatement getPark = con.prepareStatement(query);
+		getPark.setString(1, orderNum);
+		ResultSet rs = getPark.executeQuery();
+		if (rs.next()) {
+			Order tmpOrder = new Order(rs.getString(1), rs.getString(2), rs.getString(3), rs.getString(4), rs.getInt(5),
+					OrderStatus.valueOf(rs.getString(6)), OrderType.valueOf(rs.getString(7)), rs.getString(8),
+					rs.getString(9), rs.getInt(10), rs.getString(11), rs.getString(12));
+			return tmpOrder;
+		}
+		return null;
+	}
+
+	public static Order activateOrderFromWatingList(Order order)
+			throws SQLException, NumberFormatException, ParseException {
 		LocalDateTime now = LocalDateTime.now();
 		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 		LocalDateTime dateTimeOfOrder = LocalDateTime.parse(order.getDateOfOrder() + " " + order.getTimeOfOrder(),
@@ -755,7 +792,10 @@ public class MySQLConnection {
 	public static void main(String[] args) throws ParseException, InstantiationException, IllegalAccessException,
 			ClassNotFoundException, SQLException {
 		connectToDB();
+		sendSmsToActiveOrders();
+		expiredApprovedOrders();
 		sendSmsToCancelOrders();
+		
 
 	}
 
@@ -910,6 +950,8 @@ public class MySQLConnection {
 		LocalDateTime tommorow = now.plusDays(1);
 		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
 		String nowString = now.format(formatter);
+		String dateNow = nowString.split(" ")[0];
+		String timeNow = nowString.split(" ")[1];
 		String tommorowString = now.format(formatter);
 		List<Order> orders = new ArrayList<>();
 		String query = "SELECT orders.* FROM orders WHERE dateOfOrder=? AND HOUR(timeOfOrder)=? AND status='ACTIVE';";
@@ -932,7 +974,7 @@ public class MySQLConnection {
 			PreparedStatement sendSms;
 			String sendSmsQuery = "INSERT INTO smsSend (orderNum_fk,smsRecviedDate,smsRecviedTime) VALUES ";
 			for (int i = 0; i < orders.size(); i++) {
-				sendSmsQuery += "(" + orders.get(i).getOrderNum() + ",'"+nowString.split(" ")[0]+"','" + nowString.split(" ")[1] + "')";
+				sendSmsQuery += "(" + orders.get(i).getOrderNum() + ",'" + dateNow + "','" + timeNow + "')";
 				if (i != orders.size() - 1) {
 					sendSmsQuery += ",";
 				}
@@ -944,15 +986,19 @@ public class MySQLConnection {
 		return orders;
 	}
 
-	public static List<Order> sendSmsToCancelOrders() throws SQLException {
+	public static List<Order> sendSmsToCancelOrders() throws SQLException, NumberFormatException, ParseException {
 		List<Order> orders = new ArrayList<>();
-		SimpleDateFormat formatter = new SimpleDateFormat("HH:mm:ss");
-		String timeNow = formatter.format(new Date());
-		String query = "SELECT * FROM orders JOIN smsSend on orders.orderNum=smsSend.orderNum_fk WHERE "
-				+ "smsSend.smsRecviedDate=DATE(NOW()) AND ((orders.status='PENDING_FINAL_APPROVAL' AND HOUR(TIMEDIFF(?,smsSend.smsRecviedTime))>=2) OR ((orders.status='PENDING_APPROVAL_FROM_WAITING_LIST' AND HOUR(TIMEDIFF(?,smsSend.smsRecviedTime))>=1)))";
+		LocalDateTime now = LocalDateTime.now();
+		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+		String nowString = now.format(formatter);
+		String dateNow = nowString.split(" ")[0];
+		String timeNow = nowString.split(" ")[1];
+		String query = "SELECT orders.* FROM orders JOIN smsSend on orders.orderNum=smsSend.orderNum_fk WHERE "
+				+ "smsSend.smsRecviedDate=? AND ((orders.status='PENDING_FINAL_APPROVAL' AND HOUR(TIMEDIFF(?,smsSend.smsRecviedTime))>=2) OR ((orders.status='PENDING_APPROVAL_FROM_WAITING_LIST' AND HOUR(TIMEDIFF(?,smsSend.smsRecviedTime))>=1)))";
 		PreparedStatement getOrdersToSendSms = con.prepareStatement(query);
-		getOrdersToSendSms.setString(1, timeNow);
+		getOrdersToSendSms.setString(1, dateNow);
 		getOrdersToSendSms.setString(2, timeNow);
+		getOrdersToSendSms.setString(3, timeNow);
 		ResultSet rs = getOrdersToSendSms.executeQuery();
 		while (rs.next()) {
 			Order tmpOrder = new Order(rs.getString(1), rs.getString(2), rs.getString(3), rs.getString(4), rs.getInt(5),
@@ -963,7 +1009,27 @@ public class MySQLConnection {
 		if (orders.size() > 0) {
 			String queryCancelOrders = "UPDATE orders SET status='CANCELLED' WHERE ";
 			String queryDelSms = "DELETE FROM smsSend WHERE ";
+			Map<String, Map<String, List<String>>> checkWating = new LinkedHashMap<String, Map<String, List<String>>>();
 			for (int i = 0; i < orders.size(); i++) {
+				if (checkWating.containsKey(orders.get(i).getParkName())) {
+					Map<String, List<String>> dateAndTimesForPark = checkWating.get(orders.get(i).getParkName());
+					if (dateAndTimesForPark.containsKey(orders.get(i).getDateOfOrder())) {
+						List<String> timesForDateInPark = dateAndTimesForPark.get(orders.get(i).getDateOfOrder());
+						if (!timesForDateInPark.contains(orders.get(i).getTimeOfOrder())) {
+							timesForDateInPark.add(orders.get(i).getTimeOfOrder());
+						}
+					} else {
+						List<String> times = new ArrayList<String>();
+						times.add(orders.get(i).getTimeOfOrder());
+						dateAndTimesForPark.put(orders.get(i).getDateOfOrder(), times);
+					}
+				} else {
+					Map<String, List<String>> dateAndTime = new LinkedHashMap<String, List<String>>();
+					List<String> times = new ArrayList<String>();
+					times.add(orders.get(i).getTimeOfOrder());
+					dateAndTime.put(orders.get(i).getDateOfOrder(), times);
+					checkWating.put(orders.get(i).getParkName(), dateAndTime);
+				}
 				queryCancelOrders += "orderNum=" + orders.get(i).getOrderNum();
 				queryDelSms += "orderNum_fk=" + orders.get(i).getOrderNum();
 				if (i != orders.size() - 1) {
@@ -974,6 +1040,7 @@ public class MySQLConnection {
 			}
 			con.prepareStatement(queryCancelOrders).executeUpdate();
 			con.prepareStatement(queryDelSms).executeUpdate();
+			checkWatingList(checkWating);
 		}
 		return orders;
 	}
@@ -991,5 +1058,76 @@ public class MySQLConnection {
 		expirePreparedStatement.setString(3, timeNow);
 		expirePreparedStatement.setString(4, dateNow);
 		expirePreparedStatement.executeUpdate();
+	}
+
+	private static void sendSms(Order order) {
+		try {
+			LocalDateTime now = LocalDateTime.now();
+			DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+			String nowString = now.format(formatter);
+			String querySendSms = "INSERT INTO smsSend (orderNum_fk,smsRecviedDate,smsRecviedTime) VALUES (?,?,?)";
+			PreparedStatement sendSmStatement;
+
+			sendSmStatement = con.prepareStatement(querySendSms);
+
+			sendSmStatement.setString(1, order.getOrderNum());
+			sendSmStatement.setString(2, nowString.split(" ")[0]);
+			sendSmStatement.setString(3, nowString.split(" ")[1]);
+			sendSmStatement.executeUpdate();
+		} catch (SQLException e) {
+		}
+	}
+
+	private static List<Order> getWatingOrdersForParkInDateAndTime(Park park, String date, String time)
+			throws NumberFormatException, SQLException {
+		List<Order> ordersInTimeIntervalForDateInPark = new ArrayList<Order>();
+		if (park == null)
+			return null;
+		String query = "SELECT * FROM orders where parkName_fk=? AND dateOfOrder=? AND HOUR(TIMEDIFF(timeOfOrder,?))<? AND status=? ORDER BY orderCreationDate ASC";
+		System.out.println("SELECT * FROM orders where parkName_fk=" + park.getParkName() + " AND dateOfOrder='" + date
+				+ "' AND HOUR(TIMEDIFF(timeOfOrder,'" + time + "'))<" + park.getParkVisitDuration() + " AND status='"
+				+ OrderStatus.WAITING.name() + "' ORDER BY orderCreationDate ASC");
+		PreparedStatement getWatingOrdersStatement = con.prepareStatement(query);
+		getWatingOrdersStatement.setString(1, park.getParkName());
+		getWatingOrdersStatement.setString(2, date);
+		getWatingOrdersStatement.setString(3, time);
+		getWatingOrdersStatement.setInt(4, park.getParkVisitDuration());
+		getWatingOrdersStatement.setString(5, OrderStatus.WAITING.name());
+		ResultSet rs = getWatingOrdersStatement.executeQuery();
+		while (rs.next()) {
+			Order tmpOrder = new Order(rs.getString(1), rs.getString(2), rs.getString(3), rs.getString(4), rs.getInt(5),
+					OrderStatus.valueOf(rs.getString(6)), OrderType.valueOf(rs.getString(7)), rs.getString(8),
+					rs.getString(9), rs.getInt(10), rs.getString(11), rs.getString(12));
+			ordersInTimeIntervalForDateInPark.add(tmpOrder);
+			System.out.println(tmpOrder);
+		}
+
+		return ordersInTimeIntervalForDateInPark;
+	}
+
+	private static List<Order> checkWatingList(Map<String, Map<String, List<String>>> dateAndTimesToCheckForParks)
+			throws NumberFormatException, SQLException, ParseException {
+		List<Order> ordersToSendSms = new ArrayList<Order>();
+		for (String parkName : dateAndTimesToCheckForParks.keySet()) {
+			Park park = getCertainPark(parkName);
+			if (park == null)
+				continue;
+			Map<String, List<String>> dateAndTimesToCheckForPark = dateAndTimesToCheckForParks.get(parkName);
+			for (String date : dateAndTimesToCheckForPark.keySet()) {
+				List<String> timesToCheckForParkInDate = dateAndTimesToCheckForPark.get(date);
+				for (String time : timesToCheckForParkInDate) {
+					System.out.println(parkName + " " + date + " " + time);
+					List<Order> ordersInWait = getWatingOrdersForParkInDateAndTime(park, date, time);
+					for (Order order : ordersInWait) {
+						if (validateDate(order)) {
+							changeOrderStatus(order.getOrderNum(), OrderStatus.PENDING_APPROVAL_FROM_WAITING_LIST);
+							sendSms(order);
+							ordersToSendSms.add(order);
+						}
+					}
+				}
+			}
+		}
+		return ordersToSendSms;
 	}
 }
